# Use initial quads corners in ChessBoardDetector::findQuadNeighbors